backend/requirements.txt
azure-functions>=1.18.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-cosmos>=4.5.1
azure-storage-blob>=12.19.0
openai>=1.12.0
msal>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
jinja2>=3.1.3
weasyprint>=60.2
python-docx>=1.1.0
reportlab>=4.0.9
